requirements.txt
fastapi
jinja2==3.1.4
requests
typeguard==4.3.0
uvicorn
pydantic
python-dotenv